\name{covid.nssn-package}
\alias{covid.nssn-package}
\alias{covid.nssn}
\docType{package}
\title{
\packageTitle{covid.nssn}
}
\description{
\packageDescription{covid.nssn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{covid.nssn}
\packageIndices{covid.nssn}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{covid.nssn}

Maintainer: \packageMaintainer{covid.nssn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
